requirements.txt
pydantic-settings>=2.3.4
python-dotenv==1.0.0
requests==2.32.3
slack_sdk==3.33.5